+++ speech_to_text/lib/speech_to_text_web.dart
@@ -26,7 +26,8 @@ class SpeechToTextPlugin extends SpeechToTextPlatform {
   /// [debugLogging] controls whether there is detailed logging from the underlying
   /// plugins. It is off by default, usually only useful for troubleshooting issues
   /// with a paritcular OS version or device, fairly verbose
-  Future<bool> initialize({debugLogging = false}) async {
+  Future<bool> initialize(
+      {debugLogging = false, List<SpeechConfigOption> options}) async {
     return false;
   }