+++ speech_to_text/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 3.1.0
+
+### New
+  * `initialize` now supports the `options` parameter to supply platform specific options during initialization. 
+  * `SpeechToText.androidAlwaysStop` supported as an option. Forces the plugin to use the speech recognizer `stop` 
+  method even on SDK versions where that might fail. See https://github.com/csdcorp/speech_to_text/issues/150
+  * `SpeechToText.androidIntentLookup` looks up the intent name instead of using the default. This can work around 
+  some issues with security exceptions when trying to use the default. See 
+  https://github.com/csdcorp/speech_to_text/issues/153
+
 ## 3.0.1
 
 ### Fix